Apply pr review
